pyproject.toml
[tool.poetry]
name = "cli-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-ollama = "^0.2.3"
langchain = "^0.3.17"
langchain-community = "^0.3.16"
langchain-openai = "^0.3.3"
langgraph = "^0.2.69"
langgraph-checkpoint-sqlite = "^2.0.3"
scikit-learn = "^1.6.1"
beautifulsoup4 = "^4.13.0"
requests = "^2.32.3"
ipython = "^8.32.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
